Pk10/requirement.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
lxml==4.3.3
Pillow==6.0.0
PyMySQL==0.9.3
requests==2.21.0
selenium==3.141.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
